requirements.txt
flask
tensorflow
spacy
nltk
pandas
tweepy
transformer